+++ doc/source/library/architecture/architecture.rst
@@ -5,14 +5,16 @@ The internal structure of pymodbus is a bit complicated, mostly due to the mixtu
 
 The overall architecture can be viewed as:
 
+
     Client classes (interface to applications)
     mixin (interface with all requests defined as methods)
-    transaction (handles transactions and allow concurrent calls)
-    framers (add pre/post headers to make a valid package)
-    transport (handles actual transportation)
+    Lower levels are Common
 
     Server classes (interface to applications)
     datastores (handles registers/values to be returned)
+    Lower levels are Common
+
+    pdu (build/create response/request class)
     transaction (handles transactions and allow concurrent calls)
     framers (add pre/post headers to make a valid package)
     transport (handles actual transportation)
@@ -25,4 +27,3 @@ In detail the packages can viewed as:
 In detail the classes can be viewed as:
 
 .. image:: classes.png
-
