+++ pymodbus/framer/ascii.py
@@ -8,8 +8,8 @@ from __future__ import annotations
 
 from binascii import a2b_hex, b2a_hex
 
+from pymodbus.framer.base import MessageBase
 from pymodbus.logging import Log
-from pymodbus.message.base import MessageBase
 
 
 class MessageAscii(MessageBase):
