--- pymodbus/message/__init__.py
@@ -1,7 +1,0 @@
-"""Message."""
-__all__ = [
-    "Message",
-    "MessageType",
-]
-
-from pymodbus.message.message import Message, MessageType
